requirements.txt
discord.py
matplotlib
japanize-matplotlib
numpy